+++ apps/server/src/apps/deletion-console.app.ts
@@ -1,5 +1,5 @@
 /* istanbul ignore file */
 import { DeletionConsoleModule } from '@modules/deletion-console/deletion-console.app.module';
-import { runConsoleApp } from './helpers/run-bootstrap-command';
+import { runConsoleApp } from './helpers/run-console-app';
 
 void runConsoleApp(DeletionConsoleModule);
