+++ apps/server/src/apps/helpers/run-console-app.ts
