+++ apps/server/src/apps/idp-console.app.ts
@@ -1,5 +1,5 @@
 /* istanbul ignore file */
 import { IdpConsoleModule } from '@modules/idp-console/idp-console.app.module';
-import { runConsoleApp } from './helpers/run-bootstrap-command';
+import { runConsoleApp } from './helpers/run-console-app';
 
 void runConsoleApp(IdpConsoleModule);
