+++ apps/server/src/apps/server-console.app.ts
@@ -1,7 +1,7 @@
 /* istanbul ignore file */
 /* eslint-disable promise/always-return */
 import { ManagementConsoleModule } from '@modules/management/management-console.app.module';
-import { runConsoleApp } from './helpers/run-bootstrap-command';
+import { runConsoleApp } from './helpers/run-console-app';
 
 /**
  * The console is starting the application wrapped into commander.
