+++ apps/server/src/apps/sync-console.app.ts
@@ -1,5 +1,5 @@
 /* istanbul ignore file */
 import { MediaSyncConsoleAppModule } from '@modules/media-sync-console/media-sync-console.app.module';
-import { runConsoleApp } from './helpers/run-bootstrap-command';
+import { runConsoleApp } from './helpers/run-console-app';
 
 void runConsoleApp(MediaSyncConsoleAppModule);
